refactor: 액션으로 컴포넌트 분리, 단일 책임 원칙
